Refactor: 구조 수정
