+++ app/src/main/AndroidManifest.xml
@@ -8,9 +8,6 @@
 
 
     <application
-
-        android:networkSecurityConfig="@xml/network_security_config"
-        android:usesCleartextTraffic="true"
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
