+++ app/src/main/java/com/umc/upstyle/ClosetItemFragment.kt
@@ -3,17 +3,16 @@ package com.umc.upstyle
 import android.content.Context
 import android.net.Uri
 import android.os.Bundle
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.GridLayoutManager
-import com.umc.upstyle.api.ApiService
+import com.umc.upstyle.data.network.ApiService
 import com.umc.upstyle.databinding.FragmentClosetItemBinding
-import com.umc.upstyle.model.ClosetCategoryResponse
+import com.umc.upstyle.data.model.ClosetCategoryResponse
 import java.io.File
-import com.umc.upstyle.model.ClothPreview
+import com.umc.upstyle.data.model.ClothPreview
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
