+++ app/src/main/java/com/umc/upstyle/LoadCategoryFragment.kt
@@ -1,19 +1,14 @@
 package com.umc.upstyle
 
-import Item_load
-import android.content.Context
-import android.content.SharedPreferences
+import com.umc.upstyle.data.network.Item_load
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
-import androidx.core.content.edit
 import androidx.fragment.app.Fragment
-import androidx.fragment.app.FragmentTransaction
 import androidx.navigation.fragment.findNavController
 import com.bumptech.glide.Glide
-import com.umc.upstyle.databinding.ActivityClosetBinding
 import com.umc.upstyle.databinding.FragmentLoadCategoryBinding
 
 class LoadCategoryFragment : Fragment() {
