--- app/src/main/java/com/umc/upstyle/PhotoUploadActivity.kt
@@ -1,4 +1,0 @@
-package com.umc.upstyle
-
-class PhotoUploadActivity {
-}
