+++ app/src/main/java/com/umc/upstyle/data/model/ClosetCategoryResult.kt
@@ -1,0 +1,5 @@
+package com.umc.upstyle.data.model
+
+data class ClosetCategoryResult(
+    val clothPreviewList: List<ClothPreview>
+)
