+++ app/src/main/java/com/umc/upstyle/data/model/ClosetResponse.kt
@@ -1,0 +1,12 @@
+package com.umc.upstyle.data.model
+
+data class ClosetResponse(
+    val isSuccess: Boolean,
+    val code: String,
+    val message: String,
+    val result: ClosetResult
+)
+
+
+
+
