+++ app/src/main/java/com/umc/upstyle/data/model/ClosetResult.kt
@@ -1,0 +1,7 @@
+package com.umc.upstyle.data.model
+
+data class ClosetResult(
+    val userId: Long,
+    val userName: String,
+    val clothKindList: List<ClothKind>
+)
