+++ app/src/main/java/com/umc/upstyle/data/model/OOTDPreview.kt
@@ -1,0 +1,8 @@
+package com.umc.upstyle.data.model
+
+// OOTD 캘린더 미리보기 데이터 클래스
+data class OOTDPreview(
+    val id: Int,
+    val date: String,
+    val imageUrl: String
+)
