+++ app/src/main/java/com/umc/upstyle/data/model/User.kt
@@ -1,0 +1,7 @@
+package com.umc.upstyle.data.model
+
+// 사용자 데이터 클래스
+data class User(
+    val id: Int,
+    val nickname: String
+)
